refactor: update converter module
